Overwrite Descriptor write prototype
This needs to be done because React Native has no global Buffer object and so the check will cause JS to break. Assumes that a Buffer will be passed into it. It does reduce overhead. If a non-Buffer object is passed to it, the React Native bridge will at least throw an error.